refactor: 자판기가 동전을 생성하는 기능을 VendingMachineCoins 가 맡도록 변경
